+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 
 ### Chores
 
+- Updated `cardano-wallet` to revision `ffeca1d9` which includes `cardano-node` 1.20.0 ([PR 2187](https://github.com/input-output-hk/daedalus/pull/2187))
 - Set "Delegation center" as default staking screen ([PR 2185](https://github.com/input-output-hk/daedalus/pull/2185))
 - Changed the ordering of wallet addresses on the "Receive" screen so that the oldest one are on top and the newest one on the bottom of the list ([PR 2176](https://github.com/input-output-hk/daedalus/pull/2176))
 - Fixed UI issues ([PR 2152](https://github.com/input-output-hk/daedalus/pull/2152))